+++ src/api/DataSetInterface.ts
@@ -31,6 +31,11 @@ export interface IDownloadDataSetOption {
      * TRANSFER_TYPE_ASCII, TRANSFER_TYPE_BIANRY, TRANSFER_TYPE_ASCII_RDW, or TRANSFER_TYPE_BINARY_RDW defined in CoreUtils.
      */
     transferType?: string;
+
+    /**
+    * Optional encoding for download and upload of z/OS data set
+    */
+    encoding?: string;
 }
 
 export interface IUploadDataSetOption {
@@ -53,6 +58,11 @@ export interface IUploadDataSetOption {
      * TRANSFER_TYPE_ASCII or TRANSFER_TYPE_BIANRY defined in CoreUtils.
      */
     transferType?: string;
+
+    /**
+    * Optional encoding for download and upload of z/OS data set
+    */
+    encoding?: string;
 }
 
 export interface IAllocateDataSetOption {