+++ src/cli/download/data-set/DataSet.definition.ts
@@ -50,6 +50,13 @@ export const DownloadDataSetDefinition: ICommandDefinition = {
             description: "Download the variable-length data set with RECFM of V, VB, VBS, etc in rdw mode, in which " +
                 "the 4-byte RDW (Record Descriptor Word) is inserted at the begining of each record.",
             type: "boolean"
+        },
+        {
+            name: "encoding",
+            aliases: ["ec"],
+            description: "The encoding for download and upload of z/OS data set.",
+            defaultValue: null,
+            type: "string",
         }
     ],
     profile: