+++ CHANGELOG.md
@@ -3,12 +3,16 @@
 All notable changes to the z/OS FTP Plug-in for Zowe CLI will be documented in this file.
 
 ## Recent Changes
+
+- Add encoding setting in profile.
 - Support list jobs by status.
 
 ## `2.0.2`
+
 - Refine help of partitioned dataset allocation.
 
 ## `2.0.1`
+
 - Pick up `zos-node-accessor` v1.0.11 to fix listing single USS file or symbol link and update PDS dataset allocation.
 
 ## `2.0.0`