+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the z/OS FTP Plug-in for Zowe CLI will be documented in this file.
 
-## Recent Changes
+## `2.1.0`
 
 - Add encoding setting in profile.
 - Support list jobs by status.